# docs/generated/openapi-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QAgent API Gateway",
    "description": "\n## QAgent API Gateway\n\nControl-plane API for QAgent — a native Agent Runtime for long-running Agent Teams\n(plan → execute → recover → deliver) with sandboxed tools and observability.\n\n### Features\n\n- **Models Management**: Query and retrieve available AI models\n- **MCP Configuration**: Manage Model Context Protocol (MCP) server configurations\n- **Memory Management**: Access and manage global memory data for personalized conversations\n- **Skills Management**: Query and manage skills and their enabled status\n- **Artifacts**: Access thread artifacts and generated files\n- **Health Monitoring**: System health check endpoints\n\n### Architecture\n\nLangGraph is mounted in-process under `/api/langgraph` (single uvicorn process).\nThis gateway provides custom endpoints for models, MCP configuration, skills, artifacts, and more.\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/skills/{skill_name}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Skill Details",
        "description": "Retrieve detailed information about a specific skill by its name.",
        "operationId": "get_skill_api_skills__skill_name__get",
        "parameters": [
          {
            "name": "skill_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "skills"
        ],
        "summary": "Update Skill",
        "description": "Update a skill's enabled status by modifying the extensions_config.json file.",
        "operationId": "update_skill_api_skills__skill_name__put",
        "parameters": [
          {
            "name": "skill_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "skills"
        ],
        "summary": "Delete Skill",
        "description": "Permanently delete a **custom** skill by removing its directory under ``skills/custom/``. Built-in **public** skills cannot be deleted. The skill entry is removed from ``extensions_config.json`` when present.",
        "operationId": "delete_skill_api_skills__skill_name__delete",
        "parameters": [
          {
            "name": "skill_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills/install": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Install Skill",
        "description": "Install a skill from a .skill file (ZIP archive) located in the thread's user-data directory.",
        "operationId": "install_skill_api_skills_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillInstallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skills/install-from-market": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Install skill from SkillHub market",
        "description": "Download a .skill archive by slug from the public SkillHub CDN and install it into custom skills.",
        "operationId": "install_skill_from_market_api_skills_install_from_market_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillMarketInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillInstallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SkillDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the skill directory was removed"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name",
            "description": "Name of the deleted skill"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable result message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "skill_name",
          "message"
        ],
        "title": "SkillDeleteResponse",
        "description": "Response model for deleting a custom skill."
      },
...
      "SkillInstallRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "The thread ID where the .skill file is located"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Virtual path to the .skill file (e.g., mnt/user-data/outputs/my-skill.skill)"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "path"
        ],
        "title": "SkillInstallRequest",
        "description": "Request model for installing a skill from a .skill file."
      },
      "SkillInstallResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the installation was successful"
          },
          "skill_name": {
            "type": "string",
            "title": "Skill Name",
            "description": "Name of the installed skill"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Installation result message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "skill_name",
          "message"
        ],
        "title": "SkillInstallResponse",
        "description": "Response model for skill installation."
      },
      "SkillMarketInstallRequest": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Market skill slug (e.g. weather, github)"
          }
        },
        "type": "object",
        "required": [
          "slug"
        ],
        "title": "SkillMarketInstallRequest",
        "description": "Install a skill by downloading its archive from the public SkillHub market."
      },
      "SkillResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the skill"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of what the skill does"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License",
            "description": "License information"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Category of the skill (public or custom)"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether this skill is enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "category"
        ],
        "title": "SkillResponse",
        "description": "Response model for skill information."
      },
      "SkillUpdateRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether to enable or disable the skill"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SkillUpdateRequest",
        "description": "Request model for updating a skill."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}